Dash - Break down the JobCard into components; Better error states
